pyproject.toml
[project]
name = "other-programs"
version = "0"
requires-python = ">=3.12"

dependencies = [
    "beautifulsoup4",
    "cairosvg",
    "cryptography",
    "img2pdf",
    "numpy",
    "pdf2image",
    "pdfkit",
    "pillow",
    "pyinstaller",
    "pypdf",
    "pyperclip",
    "pyside6",
    "pytube",
    "pywin32 ; sys_platform == 'win32'",
    "requests",
    "lxml",
]

[project.optional-dependencies]
dev = [
    "pytesseract",
    "pytest",
]
high-fid = [
    "playwright",
]
